requirements.txt
asgiref==3.2.10
azure-common==1.1.5
azure-nspkg==3.0.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
beautifulsoup4==4.10.0
branca==0.4.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
click==8.0.2
colorama==0.4.4
comtypes==1.1.10
cryptography==3.4.8
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.2
django-allauth==0.44.0
django-cors-headers==3.7.0
django-crispy-forms==1.10.0
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.6.1
django-rest-auth==0.9.5
django-storages==1.10.1
djangorestframework==3.11.1
et-xmlfile==1.1.0
folium==0.12.1
future==0.18.2
geocoder==1.38.1
gunicorn==20.0.4
idna==3.2
inflect==5.3.0
jaraco.itertools==6.0.3
Jinja2==3.0.2
joblib==1.1.0
Markdown==3.3.4
MarkupPy==1.14
MarkupSafe==2.0.1
more-itertools==8.10.0
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
Pillow==8.0.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.1.0
pypiwin32==223
python-dateutil==2.8.2
python-decouple==3.3
python3-openid==3.2.0
pyttsx3==2.90
pytz==2020.1
pywin32==301
PyYAML==6.0
ratelim==0.1.6
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
tablib==3.1.0
urllib3==1.26.7
whitenoise==5.3.0
wikipedia==1.4.0
wincertstore==0.2
wolframalpha==4.0.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
